Split Spotable extensions into new file
